Fix fp16 overflow in real quant matmul: use fp32 accumulation
fp16 matmul of centered integers (max ~7.5) with K=1792 can produce
values up to ~94000, exceeding fp16 max (65504). Use float32 for the
integer matmul, then convert back to fp16 after.

+++ fake_quant/utils/quant_utils.py
@@ -657,15 +657,16 @@ class ActQuantWrapper(torch.nn.Module):
                 act_s_l = act_scale_l.reshape(-1, act_scale_l.shape[-1])[:, :1]
 
         # 5. Integer-equivalent matmul
+        # Use fp32 accumulation to avoid fp16 overflow (int values * K can exceed fp16 range)
         # y = sum_i (act_scale_i * w_scale_i.T) * (q_act_i @ q_w_i.T)
         N = self.W_m_int.shape[0]
-        y = (act_s_m * self.W_m_scale.T) * (q_m_flat.half() @ self.W_m_int.T)
+        y = (act_s_m * self.W_m_scale.T) * (q_m_flat.float() @ self.W_m_int.float().T).half()
 
         if q_h is not None:
-            y = y + (act_s_h * self.W_h_scale.T) * (q_h_flat.half() @ self.W_h_int.T)
+            y = y + (act_s_h * self.W_h_scale.T) * (q_h_flat.float() @ self.W_h_int.float().T).half()
 
         if q_l is not None:
-            y = y + (act_s_l * self.W_l_scale.T) * (q_l_flat.half() @ self.W_l_int.T)
+            y = y + (act_s_l * self.W_l_scale.T) * (q_l_flat.float() @ self.W_l_int.float().T).half()
 
         # Bias
         if self.module.bias is not None:
